% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covar_development}
\alias{covar_development}
\title{Example Covariates Data Frame for Development Dataset.}
\format{
A data frame with the following columns: \code{EICNo}, \code{Label}
}
\usage{
covar_development
}
\description{
A data frame containing a column for EICs (EICNo) and the corresponding labels (Label).
}
\keyword{datasets}
