% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covar_test}
\alias{covar_test}
\title{Example Covariates Data Frame for Test Dataset.}
\format{
A data frame with the following columns: \code{FileNames}, \code{SampleType}
}
\usage{
covar_test
}
\description{
A data frame containing a column for file names (FileNames) and the corresponding sample type (SampleTypes).
}
\keyword{datasets}
