% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eic_labels_test}
\alias{eic_labels_test}
\title{Example EIC Label Data Frame for Test Dataset.}
\format{
A data frame with the following columns: \code{EICNo}, \code{Label}
}
\usage{
eic_labels_test
}
\description{
A data frame containing a column for EICs (EICNo) and the corresponding labels (Label).
}
\keyword{datasets}
