% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model}
\alias{example_model}
\title{Example Trained MetaClean Model}
\format{
A list with the following slots: \code{method}, \code{modelInfo}, \code{modelType}, \code{results}, \code{pred},
\code{bestTune}, \code{all}, \code{dots}, \code{metric}, \code{control}, \code{finalModel}, \code{preProcess},
\code{trainingData}, \code{resample}, \code{resampledCM}, \code{perfNames}, \code{maximize}, \code{ylimits}, \code{times},
\code{levels}, \code{terms}, \code{coefnames}, \code{xlevels}
}
\usage{
example_model
}
\description{
A list containing the trained model information for a MetaClean classifier.
}
\keyword{datasets}
