% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fill_test}
\alias{fill_test}
\title{Example xcmsSet Object with Filled Peaks for Test Dataset.}
\format{
A list with the following slots: \code{peaks}, \code{groups}, \code{groupidx}, \code{filled}, \code{phenoData}, \code{rt},
\code{filepaths}, \code{profinfo}, \code{dataCorrection}, \code{polarity}, \code{progressInfo}, \code{progressCallback},
\code{mslevel}, \code{scanrange}, \code{.processHistory}
}
\usage{
fill_test
}
\description{
An xcmsSet object returned by the fillPeaks() xcms function.
}
\keyword{datasets}
