% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_development}
\alias{group_development}
\title{Example xcmsSet Object with Group Assignments for Development Dataset.}
\format{
A list with the following slots: \code{peaks}, \code{groups}, \code{groupidx}, \code{filled}, \code{phenoData}, \code{rt},
\code{filepaths}, \code{profinfo}, \code{dataCorrection}, \code{polarity}, \code{progressInfo}, \code{progressCallback},
\code{mslevel}, \code{scanrange}, \code{.processHistory}
}
\usage{
group_development
}
\description{
An xcmsSet object returned by the group() xcms function.
}
\keyword{datasets}
