% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xs_test}
\alias{xs_test}
\title{Example xcmsEIC Object for Test Dataset.}
\format{
A list with the following slots: \code{eic}, \code{mzrange}, \code{rtrange}, \code{rt}, \code{groupnames}
}
\usage{
xs_test
}
\description{
An xcmsEIC object returned by the getEIC() xcms function.
}
\keyword{datasets}
